rename puregetters as getters
