+++ client/app/getters/layout.coffee
