+++ client/app/getters/messages.coffee
