+++ client/app/getters/selection.coffee
