+++ client/app/getters/settings.coffee
