+++ client/test/fixtures/message.js
@@ -43,7 +43,7 @@ module.exports.createMessage = function createMessage(data) {
 
     // TODO: these tags should be tested
     // when compose feature will be implemented
-    // so as ../../app/puregetters/message.coffee
+    // so as ../../app/getters/message.coffee
     // that format messageStore for ReactComponent
     // to: [],
     // replyTo: [],
